pyproject.toml
[tool.poetry]
name = "pydantic-dynamo"
version = "0.1.5"
description = ""
authors = ["David Jetter <davidajetter@gmail.com>"]
readme = "README.md"
packages = [{include = "pydantic_dynamo"}]
homepage = "https://github.com/david-a-jetter/pydantic-dynamo"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.20.32"
pydantic = "^1.10.5"
botocore = "^1.23.32"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
black = "^23.1.0"
factory-boy = "^3.2.1"
faker = "^17.6.0"
doit = "^0.36.0"
pytest-cov = "^4.0.0"
mypy = "^1.1.1"
flake8 = "^6.0.0"
testcontainers = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100


[tool.mypy]
python_version = 3.9
plugins = "pydantic.mypy"
warn_return_any = true
warn_unused_configs = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "boto3",
    "boto3.dynamodb.conditions",
    "botocore.response",
    "factory",
    "faker",
    "testcontainers.core.container",
]
ignore_missing_imports = true